Track only points inside the domain in the piston examples
Use some small epsilon to ensure that the tracked line completely
resides within the domain and no point is outside the domain.
Previously the last point ended up outside the domain.
This was dropped silently before, but with the change in treelm
to extrapolate to points close by when outside the domain, this
setup gets confused now and runs into an error with the identified
neighbor.